[fidl][rust] Relax static bound for transports
Driver framework uses an executor with a non-'static lifetime, and so
wants to be able to spawn futures with non-'static lifetimes as well.
This change pushes the requirement that transports be 'static down to
the site where futures are actually spawned. This can increase the
verbosity for more common cases, but it should be fine.

Change-Id: I7e3893142269c3219ca92588a887a3a23ad56f9e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1217464
Reviewed-by: Ian McKellar <ianloic@google.com>
Fuchsia-Auto-Submit: David Koloski <dkoloski@google.com>
Commit-Queue: David Koloski <dkoloski@google.com>