[timekeeper][tests] Remove invalid assertions.
The timekeeper tests seem to have been making a few assumptions about
how a kernel-clock's generation counter works, which are not currently
guaranteed by public documentation.

Specifically, it appears to have been assuming that the generation
counter for a clock (accessible via a get_details operation) would
always start at 0, and always increment by 2 with every successful
update.

While the public docs are limited when discussing the behavior of the
generation counter, they do not guarantee this behavior.  In fact, it
does not seem to guarantee all that much at all.  Mostly, it implies
that the "generation nonce" can be used to assist in "Know[ing]
whether the clock transformation has been changed since the last
zx_clock_get_details() operation".

The implication of this is that the strongest behavior we can
guarantee is "if the transformation has changed since our last
observation made with get_details, that the generation nonce will be
different".  It does not say that the value starts at zero, counts up
by 2, or even counts up at all.

So, remove these assumptions from the Rust tests.  They could be
reworked to check for there being a difference, but it does not seem
necessary given that there are already dedicated clock unit tests
which verify the generation counter behavior.

Also, update the documentation to make these guarantees explicit
instead of simply implying them.

Change-Id: I03575a88fddeac877495e4a7ebf581cdfffd6ca7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1222505
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Filip Filmar <fmil@google.com>