[bazel_sdk] remove generated constraints from rules_fuchsia
This CL moves the generated API level constraints out of the static
rules_fuchsia repository and into the generated fuchsia_sdk repository.
These constraints are moved from //fuchsia/constraints:* to
//constraints:* to separate them from the constraints that are
statically defined in the rules_fuchsia repository. The generated rules
and other usages are updated as part of this CL. None of the OOT
repositories are using these same constraints so we do not need to worry
about their usages, however, if they do want to use them they can
reference them via the @fuchsia_sdk//constraints:* path.

This CL does not move the CPU related constraints from rules_fuchsia
because they are static and do not depend on the contents of the IDK.

Fixes: 401038394
Change-Id: Idf8a3bf4dba5db58aa659c0fc8baf0ca19d33dfe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1221246
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>