[ui] Configure scene_manager with thread & memory profile.
On a heavily contended system it can a long time for scene_manager
to get scheduled and to resolve page faults when it needs to handle
input events. This will allow product integrators to tune its priority
relative to other programs contending for resources.

This should be a no-op on products which do not configure a role.

Bug: 397790128
Change-Id: I0b303f8aad116742f4f93e6f1c9a6860fea8cee2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1221090
Reviewed-by: Caroline Liu <carolineliu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>