[fuchsia-controller] Use IntEnum instead of IntFlag of enum types
The dynamic FIDL bindings now use IntEnum instead of IntFlag for enum
types. This makes more sense because the values of a FIDL enum are not
intended to be logically combined the way values of a FIDL bits are.

Bug: 355717613
Change-Id: I1c80e007e14dabf88b7f40b79229aa5c1e39ad76
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1206926
Reviewed-by: Andrew Davies <awdavies@google.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>