[bazel_sdk] stop merging rules into fuchsia_sdk
This change removes the logic that merges the starlark rules into the
fuchsia_sdk repository. The users of the SDK have been migrated to the
new way of loading rules from rules_fuchsia.

There is still some cleanup work to be done with the build scripts to
remove the use_rules_fuchsia argument but that will be done in a later
CL since they are just setting that value to True everywhere and the
call sites for the generators have not been changed.

Bug: 381104603
Change-Id: Ibfd8fc3d2215e29694fc7d482f4d83ec40a211e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1222106
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: David Fang <fangism@google.com>