[fdf][rust] Use an `Arc` in with_raw_dispatcher
A lot of contortions in tests happen because of the non-'static
test dispatcher. The reason for making it non-'static was to
prevent it from escaping the test, but it's not really worth it for what
we're doing in most of the tests.

Instead, give the test an `Arc` of the dispatcher and then make sure
that there's only one outstanding reference to it after the test body
has run.

Multiply: fdf_rust
Change-Id: I0bb0bd112d23ef6dffd3ce72c0cfa6d12a8649a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1222022
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>