[fdf][rust] Tasks have to be static, unfortunately
It was incorrect to allow clients to post tasks to the driver runtime
dispatcher that are non-'static, because it's fairly trivial to create
situations where the task borrows something that lives less time. The
only reason it appeared like this should work was because we end up
erasing lifetimes in the implementation so the compiler couldn't tell us
it was wrong.

Multiply: fdf_rust
Change-Id: I94a5862147d4cde8634b20547c9ef693e83c0388
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1222023
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>