[starnix][range_map] Implement RangeMap2
This CL contains a functioning copy-on-write implementation of RangeMap
that passes all the existing RangeMap tests. This code is an evolution
of the CowMap code, but specialized to implementing a range-based map.
Removing the abstraction between the RangeMap and the underlying Map
opens the door for future optimizations, such as removing the O(n^2)
gap search.

Change-Id: Ibe3c5d09e76448a17e5002fe6f6d16073ac361f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1224745
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>