[netstack3] ResourceCounterContext non-mut ctx
Update the bounds on the ResourceCounterContext so that it operates on
an immutable reference, rather than a mutable reference.

In doing so, `increment` is renamed to `increment_both`. Now that it
takes an immutable self, the compiler had a hard time determining which
`increment` was being called (either the one provided by
`CounterContext` or `ResourceCounterContext`). Giving it a separate name
disambiguates things.

Change-Id: Ia30d5f426b7317fc1165d6ececde06a775c92472
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1224384
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>