[netstack3] Simplify counter context traits
Remove the `with_` style methods from `CounterContext` and
`ResourceCounterContext`, replacing it with a flat method that returns a
reference to the underlying counters. This simplification makes it
possible for a follow up CL to access both "stack-wide" and
"per-resource" counters at once.

I suspect the original pattern was in place to facilitate putting
counters behind locks. However, that ability is not used by any of our
code today.

Change-Id: Ic0402885fe2f954b6f4d910975508435c9f7af8c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1225424
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>