[build] Assert PB exists with product.board
Fuchsia release infrastructure currently requires that a PB exists
with the product.board name.

We do not enable this assert on multi-product builds, because the
product.board are generic and not expected to be equal to a PB name.

Bug: 402187053, 401966851
Change-Id: I67f13647697dba6e9b7a5df1ae5fe15ecc5120d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1224546
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Mike Comfoltey <comfoltey@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>