[bazel] canonicalize label in refresh-bazel-compdb
The refresh-bazel-compdb script allows a user to pass in a label. This
label must point to a fuchsia_package but the fuchsia_package target
adds a suffix to the name to make it work with the workflow tasks. This
can confuse the user so we removed the need to pass in this suffix.

Fixes: 402509201
Change-Id: Ic6100befa4d12099a0c739eef0ad1b82f915f5a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1225871
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>