[firmware] cipy_utils.py out-of-tree support.
cipd_utils.py contains some functionality that we will commonly want in
our firmware repos; the easiest way to do this will be to support
running out-of-tree and copying it around as-needed.

In this CL, we allow the `cipd` utility to be on $PATH instead of
requiring a Fuchsia checkout, and support an `--allow-dirty` flag to
support generating manifests from a dirty repo, which is useful when
developing and testing release scripts out-of-tree.

Unrelated, this CL also logs stdout/stderr when a subprocess fails,
which makes diagnosing errors much easier.

Bug: b/393584187
Change-Id: I24bcdbae7a33fb3c6127fc7f9b355fe48635d5fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1226664
Fuchsia-Auto-Submit: David Pursell <dpursell@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yecheng Zhao <zyecheng@google.com>