[assembly] Switch scrutiny structured config test to use assembly's example aib instead of core shard
This is the last user of the core shards and core realm definition in
the legacy bundle.  So switch it to a component that's part of the
example / testing AIB, so that the core_shard() and core_realm_definition()
templates can be removed.

Bug: b/385412024
Change-Id: I9dbf5ef94677e86670c3fbaacc5c6939bda7ba9b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1178716
Reviewed-by: Laura Peskin <pesk@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>