[sestarnix] Implement and call bpf_map_alloc hook.
`bpf_map_alloc()` now assigns a security context to the BPF map and it
is called by the `sys_bpf()` syscall.
In this CL we also introduce a wrapper for the ebpf_api::PinnedMap, so
that we can store the BPF Map's security state.

Note that the security state itself is for now unused. It will be used
by the bnf_* hooks in follow-up CLs.

Bug: 364570322

Change-Id: I0363000f5b24ef8ae698c4779e476b34bd9821fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1224304
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Nelly Vouzoukidou <nellyv@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>