[build] Enable linker gc for Rust targets
This patch adds the necessary compiler and linker flags
to enable linker gc for Rust targets. There are some
symbols that are undefined in unused functions, and this
patch lets the compiler and linker to be able remove the
unused functions and avoid the undefined symbol error in
the linker.

Bug: 402443020, 390250858
Change-Id: I1084275b12d2159b6ddd0126649db6a1ce24a519
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1226604
Fuchsia-Auto-Submit: Gulfem Savrun Yeniceri <gulfem@google.com>
Commit-Queue: Gulfem Savrun Yeniceri <gulfem@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>