[kernel][vm] Free Loaned Pages Holder
Instead of moving loaned pages directly from a VMO->PMM they instead go
via a holding phase. While in this holding phase they are considered
immutable, and it is possible to wait for the page to cease being in
this temporary holding object.

An immutable holding object is necessary as the fine grained
VMO locking needs to be able to drop the lock to complete a range
change update, but a range change update must be completed before the
pages can be returned to the PMM.

Bug: 338300943
Change-Id: Ibed216283904e7382e5a508c050dc65df0dab1c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216964
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>