[tee][binder-proxy] Access shared memory region using virtio-pmem
This routes the fuchsia.hardware.virtio.pmem service to binder-proxy in
the tee realm and adds a configuration 'use_virtio_pmem' to
binder-proxy. If this configuration value is set to true, binder-proxy
waits for a device to be published in the pmem service and collects a
VMO from the service. The VMO is not yet used for anything.

Bug: 397758089

Change-Id: I4abd8ad9696518023a841b60bea2ec7c3ff468bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1216548
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: James Robinson <jamesr@google.com>