[sestarnix] Implement and call bpf_prog_alloc hook.
Similarly to `bpf_map_alloc()`, the `bpf_prog_alloc()` hook assigns
the security to the BPF prog. The hook is called from the sys_bpf()
syscall.

Note that the security state of the Program, is not accessed right
now, but it will in a follow-up CL.

Bug: 364569869
Change-Id: I7dc903c02b7ec132c88ed3f1bcc8b0926d76a8c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1225207
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Nelly Vouzoukidou <nellyv@google.com>
Reviewed-by: Ambre Williams <ambre@google.com>