[debugger] Add ElfMemoryRegion
This provides a way to directly access raw ELF files, decompressing
sections as needed. This is particularly useful for the unwinder which
expects to be able to read ELF data directly from a Memory
implementation (which is just a simple opaque reading API).

This CL doesn't hook up any of the users of this class, that will be
done in followups.

Bug: https://fxbug.dev/292107156
Change-Id: I1d4e72137d45cef40926cd61ad61e8a6bc714147
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1195606
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>