[tools][fuchsia-sdk-run] Removing fuchsia-sdk-run
This tool was intended for use OOT when switching between multiple
Bazel SDK projects. However, the project has shifted priorities and
OOT multiple projects are not going to happen as hoped.

This tool imposes some constraints on the design and structure of the
IDK and SDK which are causing friction to change.

When the need arises for a tool like this again, it may be restored,
time will tell....

Bug: 386416011
Change-Id: Ieb8dc48b48d6f9c7efca641c444feb7b3590ea38
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1227064
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>