[ffxutil] support strict config check-ssh-keys
Support ffx --strict config check-ssh-keys

This required a bit of a refactor, because the handling of ssh keys had
cyclical dependencies: some clients of FFXInstance wanted to provide
ssh key paths when constructing the instance, others wanted to query the
paths from the FFXInstance. There are also some clients which know both
the public and private key path, whereas others know only the public
one.

`ffx --strict config check-ssh-keys` requires both to be specified.
The current logic this requires the keys to be passed when in strict
mode (which is only used by botanist currently); when not in
strict mode, the keys are allowed to be determined from the environment.

Bug: 400798715
Change-Id: I1f201178e43825662f284be6260995105fb3cd3a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1224748
Reviewed-by: Peter DeLong <pdelong@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Reviewed-by: Gulfem Savrun Yeniceri <gulfem@google.com>