[bazel] workspace_utils.py: Fixed missing tracked files
The generate_fuchsia_workspace() function returned a set of
file paths that was missing a number of important input files
that are read by the function. As a consequence modifying any
of them would not properly trigger a regenerator invocation
on the next `fx bazel` or `fx build` call:

- build/bazel/templates/template.bazelrc
- build/bazel/templates/template.bazel.sh
- build/bazel/templates/template.platform.mapping

This fixes the issue by:

- Adding a `read_text_file()` function and `input_files`
  property to the GeneratedWorkspaceFiles class.

- Using them to read input files in record_fuchsia_workspace()

- Add the content of generated.input_files in the result
  of generate_fuchsia_workspace().

Manual comparison of the `build.ninja.d` file generated before
and after the CL confirms that the missing input files are
now properly listed in that file.

Change-Id: Ifd05f1ba4802c6dd34367a5094d52f690b09d3dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1227466
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: David Turner <digit@google.com>