[block_server] Fix FIFO race
fuchsia_async::Fifo is not intended to be used with multiple writer
tasks.  (See linked bug for details).  Therefore, we must serialize
write access.

The old FIFO loop worked as follows:
- Read a batch of requests from the FIFO.
- For each request, spawn a task to:
  - Process the request, and
  - Write the response onto the FIFO.

The last step is the problematic one: each of these tasks might
concurrently write to the FIFO.

To fix this, the FIFO loop was changed to use an event-driven loop.  The
new loop logic is:
- Do *one* of the following, in priority order:
  - Send finished responses on the FIFO.
  - Make progress on processing active requests.
  - Receive new requests on the FIFO.
- If any new requests were received, spawn a processing task for each.

This ensures that the FIFO is only being read to or written from
one-at-a-time.  It also provides natural flow control, as we will always
prioritize responding to requests over receiving new ones.

I suspect this is not the best arrangement for raw throughput, but we
can revisit this later as needed.  It should be noted that this
interface will not be used on most configurations due to passthrough
I/O; we only use this on configurations where the rust drivers must be
involved (e.g. FVM+zxcrypt).

Bug: 401250253
Change-Id: I4be10e9f507caacd59be57d9005bd1cb20ffab8a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1224444
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>