[vscode] Add trailing slash on remapPrefix
In some cases vscode seems to be inconsistent about whether or not the
${workspaceFolder} substitution has a trailing slash leading to things
only working intermittently or even not at all. This may have been made
worse some time recently as well.

This adds a trailing slash to the prefix so that it should always have
it, and seems to quell the errors in the log like:

```
File with cargo diagnostic not found in VFS: file not found: /usr/local/google/home/mgnb/fuchsiasrc/...
```

Change-Id: I854b592e967f18eb589a70677f65e047cc2fe036
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1230464
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>