[paver] Clone pager payload
Instead of attempting to assert that the payload VMO is not a root
pager VMO, this just directly creates a clone. Presently the check does
not work since the payload_vmo could be a reference clone, which is
both pager backed and has a parent koid, and so passes the check, but
since it still is the root VMO it could suffer from eviction.

Change-Id: I83c5cc3c79859c94e517f9a043d8bf1594538998
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1228604
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>