[brcmfmac] Adjust max TX glom frames
Recently the MTU in the driver was increased from 1500 to 1514, along
with that products are moving to ns3 which is better at utilizing frame
space. This resulted in the total size of a TX glom batch exceeding the
unsigned 16-bit limit in the glomming protocol header.

Adjust the maximum number of frames transmitted at a time to match the
maximum number of frames that can be TX glommed. Also add an error
message to indicate if the total size exceeds the maximum value. In the
future, this should ideally be improved to better recover if this
happens.

Bug: 402604249
Test: Manual testing
Change-Id: I80e2290bf6435468ad7055abeef33f0cbaa6c53e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1228148
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>