[zxio] Update zxio_create() on unknown type to successfully wrap
When zxio_create() is passed a valid handle to an object whose type it
does not understand it wraps the handle in a "holder" object that can
close the handle or provide it back to a caller. This allows zxio to
wrap any sort of handle in a zxio object. This change updates the return
value of this case to ZX_OK to make it easier for callers to understand
that the object is still valid. This handle holder behavior is useful
for transferring handles via zxio objects in some APIs - for example to
wrap an arbitrary handle in an fd to pass it through a binder protocol.

Previously this case returned ZX_ERR_NOT_SUPPORTED. This error case was
used during the fdio to zxio transition so that fdio could first ask
zxio to wrap a handle and if it was not something that zxio understood
fdio could then apply its own wrapping logic. That transition was
completed some time ago and so this confusing error handling is no
longer needed.

Change-Id: I449758773928b46b71c624d90bc0ee0a027fb983
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1231204
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>