[honeydew][python] Fix coverage.sh script.
This allows for fuchsia-controller to be imported using the system
search paths in addition to the current shared lib directory.

This also updates the paths used in the install.sh and coverage.sh
scripts in use with Honeydew so that they are able to locate the updated
`fuchsia_controller` library.

Fixed: 394153187

Change-Id: I6b1f3064067333f6c597024d81fcad418d21b3ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1228985
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Reviewed-by: Charles Celerier <chcl@google.com>
Reviewed-by: Pavan Kumar Juturu <jpavankumar@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>