[ffx][emu] Only generate a vsock CID if config value is 0
Right now we unconditionally overwrite whatever value is provided in the
config. While the user in most cases wouldn't care what value ends up
being used, it's strange to have this behaviour while exposing it as a
knob in the config. Accordingly, we update the default as 0 (an invalid
CID) and only overwrite it in that case. If any non-zero value is
provided, then that's the user intended and that's what the user will
get.

Change-Id: I7e129bc7453d5b5f13787624720f5ddeec9f27c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1227674
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>