[starnix][cgroup] Freeze newly created Task in frozen cgroup
As a freeze operation is propagating, `clone_task()` can happen
simultaneously, and the newly cloned task can "escape" from the freezing
cgroup, allowing it to be scheduled. This change fixes this by:

1. Locking the Cgroup hierarchy's PidTable, so that processes cannot
   move between cgroups during Task creation.
2. Initializing the new Task with an existing KernelSignal::Freeze.

Verifying this behavior in a test is hard, because we cannot guarantee
the ordering between a `fork()` and an ongoing `freeze()`, and depending
on which happens first, the newly spawned `Task` can both be scheduled
or never scheduled.

Added stub log for CLONE_INTO_CGROUP, it is not implemented in this
change.

Bug: 402530518
Multiply: cgroup_freezer_test
Change-Id: Ie484a4757289dd74fdab9842353130add5c07da1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1230023
Commit-Queue: David Song <wintermelons@google.com>
Reviewed-by: Guocheng Wei <guochengwei@google.com>