[fdf] Fix issue with disabling a nested composite
This fixes an issue where a composite that is colocated inside another
composite could not be disabled fully as the composite rematch was not
being performed. On the driver index, the disable handler does this, so
here I've added the same into the driver manager after it receives
confirmation from the driver index.

Change-Id: I78d0778b5773949e7fcc3af6e66bfe9dc9cd4c56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1221227
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>
API-Review: Suraj Malhotra <surajmalhotra@google.com>