[kernel] Create singleton KTrace class
This CL modernizes the ktrace library API by introducing a new
singleton class called KTrace that encapsulates the KTraceState.
The Emit* functions are now static methods of the KTrace class.

Because this is a refactor, it should be a functional no-op.

Bug: 42069955
Change-Id: I30cecc1376394b9524544827bad7c26360c2311d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1231580
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>