[devicetree-boot-shim] Fix DevicetreeSerialNumberItem
By mistake we are immediately add the offset of the cmdline option after
look up, before validating that it was found. This leads to two possible
issues:

(1) We can loop around and provide invalid serial number in the fallback
case.

(2) If the boot args are small enough, the offset will fall out of range
leading to a crash.

Fixed: 403614777
Change-Id: I085b440dadeaf5ccf78eb7e83b67fa76e8c41118
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1230224
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>