[kernel] Ensure .dynamic ends up as read-only
The section may statically be writable, but in any case after relocation
it should be read-only. This change ensures that by updating the layout
to place .dynamic in rodata if it was originally read-only or in RELRO
if writable.

This fixes a regression with BFD on arm64 and x86 where .dynamic was
writable, leading to a writable rodata segment.

Bug: 42164859
Change-Id: Ibf66b7de68409ff3b3d288ed65e49c822cb8aba0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1232064
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>