[kernel] Move ktrace_init into KTrace singleton
This moves ktrace_init, which is the LK_INIT_HOOK that initializes
the tracing subsystem in the kernel, to a static method in the
KTrace singleton. Some of the operations in the init function have
been reordered: the early exit when ktrace is disabled now runs
before the RegisterCallback and the CategoryBits are set up, since
they will not be needed if ktrace is unused.

This is the final callsite that made use of the public accessor for
the internal KTraceState. With its removal, we can finally make that
accessor private and prevent any files that include ktrace.h from
being able to access that internal state.

Bug: 42069955
Change-Id: I0b66cfd2eee774ca38752eb91606c976c45cbb4f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1232105
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>