[fidlgen_go] Treat flexible methods/events as strict
To assist with ongoing FIDL protocol migrations, we allow fidlgen_go to
generate code for handling flexible methods/events by treating them as
strict. All remaining Go code is used to communicate with in-tree
servers, so we always know the exact set of methods which
servers/clients on both sides need to support.

This is required to allow us to transition certain io1 methods to be
flexible to allow a graceful phase-out. This will allow us to complete
the ongoing fuchsia.io migration without being blocked on the Netstack2
removal.

Bug: 400784122
Change-Id: I71405bf0e834c0ab15d67496e445646c4b3b8926
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1225704
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>