[botanist][qemu] Removing //tools/qemu from botanist
This removes the dependency from botanist on //tools/qemu.

Botanist has been migrated to use ffx to manage emulator instances, so
this code is not used.

There was a mapping of cpu architecture names that was being referenced,
that are now copied into emulator.go.

This change is being done to clean up tech debt and more clearly
understand what code is being used and what can be cleaned up and
replaced with more modern or heavily used libraries.

Change-Id: I7095cddeec142038fb40948bf4f4499917e4af5b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1232108
Reviewed-by: Ina Huh <ihuh@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>