[ffx] track remote overnet_id in the HostPipe
If we get a remote overnet id from the host-pipe, we'll need to
track it, and return it to callers. The specific plan is for
each TargetHandle object to request the overnet_id, and use that
to determine whether it should redirect its requests to a previously-
existing target.

Bug: 395632395
Change-Id: Id563882d65c7bb5f932e3fc560e9cf495d274e2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1229246
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>