[ffx] reuse Targets when overnet node is shared
When a FIDL Target handler makes an RCS connection, check if it
is the same overnet id as one that already exists. If so, redirect
further Target requests to the previous target.

This required adding the ability to asynchronously wait for the
remote overnet id to arrive.

Bug: 395632395
Change-Id: Iaf91b4bd3dfcaa79bae9e3bb3dcf62e97cd8c17e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1229247
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>