[component_manager] remove races from events integration tests
An event stream will only receive events starting at the point the event
stream connection is established. Some integration tests of this had
eager components in their realm, would connect to an event stream, and
the assert that they could see the resolve and start events of these
eager components. This is inherently racy, and the race is more likely
after upcoming refactors.

This commit reworks these tests such that the components whose events we
want to observe only start after the event stream is connected to.

Change-Id: I46fda269abe1a31985c930cb9a2bf61551c8dead
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1231864
Reviewed-by: Amanda Tait <atait@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>