[blobfs] Support for blob creator AllowExisting flag.
Supports AllowExisting flag so that a blob writer holds a reference to
a blob that it expects to overwrite. During commit it replaces the
loader info for the old blob with the new one and commits the old blob's
removal to the journal in the same transaction.

If the disk writes fail such that the transaction cannot be completed,
then the blob is marked as corrupt to prevent attempts to load it
afterward as the metadata and/or block data for the blob is in an
unknown state.

Bug: 400777068

Change-Id: I6f4e0b04abda709abf0a6bee6cb9e2a16b509e94
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1229252
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>