[ffx][trace] Refactor trace post-processing
This is a refactor that should have equivalent behavior, but that will
make the next changes I need to make to IPC trace symbolization simpler.
Specifically, to correctly generate round-trip flow messages for async
two-way FIDL calls we need a two-pass approach.

This moves from having a TraceProcessor struct that you set up with some
state and then consume with its run method, just have a
process_trace_file function that takes all of the state and
configuration and returns the result. This makes the lifetime of some of
the state and configuration simpler to reason about.

Along the way I added a couple of helpers to abstract away some of the
trace file handling.

First a TraceIterator that wraps the SessionParser to provide an
iterator yielding the parsed and raw form of each record in the trace.

Secondly the process_event_records function that parses a trace buffer
(using TraceIterator) and passes every record that is an event through a
callback function. This function can manipulate the contents of the
record and return a revised value.

Change-Id: I04e4e1081cd1a448fa72475d5712deb13f504307
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1230508
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>